UI inconsistencies fixed pt2
